Add code " new basket", new product, new article

+++ app/src/main/java/com/example/shopping_list/navigation/NavControllerApp.kt
@@ -12,6 +12,6 @@ fun NavHostController.navigateToScreen(route: String) =
         }
     }
 
-fun NavHostController.navigateToProducts(basketId: Int) {
+fun NavHostController.navigateToProducts(basketId: Long) {
     this.navigateToScreen("${ProductsBasket.route}/$basketId")
 }
